api/requirements.txt
boto==2.48
botocore==1.8.40
dj-database-url==0.4.2
Django==1.11.10
django-cors-headers==2.0.2
django-filter==1.1.0
django-ses==0.8.5
django-storages==1.6.5
djangorestframework==3.7.7
gunicorn==19.7.1
psycopg2==2.7.4
#pyOpenSSL==17.2.0
stripe==1.77.2
